Fix two clang-tidy warnings

+++ src/framebufferimpl.cpp
@@ -13,7 +13,8 @@
 namespace jngl {
 
 FrameBufferImpl::FrameBufferImpl(int width, int height)
-: height(height), texture(width, height, width, height, nullptr),
+: height(height),
+  texture(static_cast<float>(width), static_cast<float>(height), width, height, nullptr),
   letterboxing(glIsEnabled(GL_SCISSOR_TEST)) {
 	GLint tmp;
 	glGetIntegerv(GL_FRAMEBUFFER_BINDING, &tmp);
